requirements.txt
requests
python-dotenv
faiss-cpu
numpy
pandas
matplotlib
